requirements.txt
apache-airflow
apache-airflow-providers-apache-spark
pykafka
pyspark
apache-airflow
